move to streams.python package
